GenAI/groq/requirements.txt
langchain_openai
langchain_core
streamlit
langchain_community
python-dotenv
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
wikipedia
arxiv
langchainhub
langchain-groq